+++ src/sequences/SequenceArray/SequenceArray.hpp
@@ -11,12 +11,12 @@ class SequenceArray : public Sequence<T>{
     
     SequenceArray<T>* appendInternal(T item);
     SequenceArray<T>* prependInternal(T item);
-    SequenceArray<T>* insertAtInternal(T item, int index);
+    SequenceArray<T>* insertAtInternal(T item, size_t index);
     
     public:
 
     SequenceArray();
-    SequenceArray(T* item, int count);
+    SequenceArray(T* item, size_t count);
     SequenceArray(const SequenceArray<T>& other);
     
     virtual SequenceArray<T>* Instance() = 0;
@@ -30,13 +30,13 @@ class SequenceArray : public Sequence<T>{
 
     T GetFirst() const override;
     T GetLast() const override;
-    T Get(int index) const override;
+    T Get(size_t index) const override;
     size_t GetLength() const override;
-    Sequence<T>* GetSubsequence(int startIndex, int endIndex) const override;
+    Sequence<T>* GetSubsequence(size_t startIndex, size_t endIndex) const override;
     
     Sequence<T>* append(T value) override;
     Sequence<T>* prepend(T value) override;
-    Sequence<T>* insertAt(T item, int index) override;
+    Sequence<T>* insertAt(T item, size_t index) override;
     Sequence<T>* Concat(Sequence<T> *list) const override;
 };
